+++ test/integration/Entity/TestVariableStorableTest.php
@@ -19,28 +19,28 @@
  *
  */
 
-namespace oat\taoResultServer\test\Entity;
+namespace oat\taoResultServer\test\integration\Entity;
 
-use oat\taoResultServer\models\Entity\ItemVariableStorable;
+use oat\taoResultServer\models\Entity\TestVariableStorable;
+use taoResultServer_models_classes_OutcomeVariable;
 
-class ItemVariableStorableTest extends \PHPUnit_Framework_TestCase
+class TestVariableStorableTest extends \PHPUnit_Framework_TestCase
 {
-
     public function testSerializeAsExpected()
     {
-        $resultVariable = $this->getMockBuilder(\taoResultServer_models_classes_Variable::class)->disableOriginalConstructor()->getMock();
+        $resultVariable = $this->getMockBuilder(taoResultServer_models_classes_OutcomeVariable::class)->disableOriginalConstructor()->getMock();
 
-        $var = new \oat\taoResultServer\models\Entity\ItemVariableStorable('deliveryResultIdentifier', 'test', $resultVariable, 'item', 'callIdItem');
+        $var = new TestVariableStorable('deliveryResultIdentifier', 'test', $resultVariable, 'callIdTest');
 
         $this->assertInstanceOf(\JsonSerializable::class, $var);
         $this->assertEquals([
             "deliveryResultIdentifier" => 'deliveryResultIdentifier',
             "test" => 'test',
-            "item" => 'item',
+            "item" => null,
             "variable" => serialize($resultVariable),
-            "callIdItem" => 'callIdItem',
-            "callIdTest" => null,
-            "uri" => 'deliveryResultIdentifiercallIdItem',
+            "callIdItem" => null,
+            "uri" => 'deliveryResultIdentifiercallIdTest',
+            "callIdTest" => 'callIdTest',
             "class" => get_class($resultVariable)
 
         ], $var->jsonSerialize());
