+++ test/unit/RestResultServerTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace oat\taoResultServer\test;
+namespace oat\taoResultServer\test\unit;
 
 use oat\tao\test\RestTestCase;
 use oat\taoResultServer\models\classes\ResultServerService;
